+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v0.1.8
+* Fixed: Madness didn't reduce cost of Weaponized cards
 
 ### v0.1.7
 * Changed Auto-Strike: now deals 3(5) damage when drawn, instead of 3(4).